Give hubot 12 seconds to start up.
Supervisord considers any process which exits after one second or
longer (by default) to have successfully started, and will try
to restart it without any penalty.

It can take a second or two for hubot to start up and open network
connections and auth and so forth, so give us a little more breathing
room.