dropwizard-api: Add proccess supervisor for APIs
ECSOPS-20

This commit changes the dropwizard-api role to set up the server for
automated deployment. To wit, it sets up an instance of supervisord to
supervise the API processses and installs some scripts to help manage
them.

It also remove the tasks for cloning an API repo,
since jenkins will be responsible for copying the API jars into place.