pip: fix installation under CentOS 6.8
ECSOPS-109

It seems the EPEL repository wasn't getting configured correctly on
CentOS 6.8 for whatever reason, causing the python-pip install to fail
because the package could not be found. The epel-release package seems
to exist on 6.x, so just use it instead of attempting to add the EPEL
repository manually. Not sure why we were only using it for 7.x.

Tested on CentOS 6.7 and 6.8.