openjdk-8: update ca-certificates-java when installing on Debian Jessie
ECSPCS-116

The latest version of openjdk-8 from jessie-backports is
incompatible with outdated version of ca-certificates-java which
comes with Debian Jessie, so we need to install an updated version
from jessie-backports.

We need to specify state=latest or else the package will not be
updated, since an old version is already installed. We need to
specify default_release=jessie-backports or else the package will
not be updated, since apt prefers packages from the main repo and
will only install packages from the backports repo if explicitly
asked.