dropwizard-api: read environment variables before launching
ECSOPS-67

ECSOPS-103 created a common repo for config files. In order to keep
credentials off of github, it was decided to set them with environment
variables. This commit adds support for loading environment variables
from a file before launching an API.

If the startup script finds a file named environment.env in the
api directory /apis/apis/$name, then it will scan it for lines of the
form NAME=value and export them to the environment (NAME must start
with an uppercase character).