Check docker-ce version before changing versionlocks
Previously, yum versionlocks for docker-ce would be removed and added
every time the Docker role was run. ansible-lint would raise a warning
since the command module was used without any conditions. To fix this,
the docker-ce version will be checked and versionlocks and installation
will only occur if a different version is currently installed.