disable xpack
xpack required a license, and tons of exceptions were being thrown in elasticsearch logs about the fact we were using an expired license. Furthermore, it was causing the locations frontend healthcheck to fail because I guess if we're using an expired license for a plugin we aren't even using, that means the datasource is unhealthy

therefore, I disabled xpack. Good riddance :)